+++ CHANGES.md
@@ -30,8 +30,12 @@ Update version in extension.json.
 
 ## Versions
 
-### 0.1.11-SNAPSHOT
+### 0.1.12-SNAPSHOT
 YYYY-MM-DD
+Speechoid: 0.1.2
+
+### 0.1.11
+2025-05-13
 Speechoid: 0.1.2
 
 * [T311047](https://phabricator.wikimedia.org/T311047) Add a view link to edit lexicon entries as raw text